feat(events): add schema.org structured data and extract shared EventLandingPage component
fixes #651 and #653
