+++ src/components/EventLandingPage.js
@@ -1,0 +1,30 @@
+import Layout from "@theme/Layout";
+import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
+import EventLanding from "@site/src/components/EventLanding";
+import events from "@site/src/data/events";
+
+export default function EventLandingPage({ slug }) {
+  const { i18n } = useDocusaurusContext();
+  const isZh = i18n.currentLocale.startsWith("zh");
+  const event = events.find((e) => e.slug === slug);
+
+  if (!event) {
+    return (
+      <Layout title={isZh ? "未找到活动" : "Event not found"}>
+        <main className="container margin-vert--xl">
+          <h1>{isZh ? "未找到活动" : "Event not found"}</h1>
+        </main>
+      </Layout>
+    );
+  }
+
+  return (
+    <Layout
+      title={isZh ? event.title.zh : event.title.en}
+      description={isZh ? event.description.zh : event.description.en}
+      image={event.banner}
+    >
+      <EventLanding event={event} />
+    </Layout>
+  );
+}
